organize physics
